Changed the 1 second delay in the disconnect LoopingCall to a variable.
This allows to modify it in special cases (for example in tests) and
set it to a lower values when we don't want to a full second.

As disconnection normally happens quite quickly, its value was also
lowered to .5 seconds. Additionally the LoopingCall now init argument
was set to True, in order to execute the first LoopingCall iteration
straight away. This allows to fire the deferred as quickly as possible
in the normal case and to wait the given delay only if the
disconnection takes longer as expected.