Add new connection implementation in preparation for replicaSet failover.
This is the first step in a long process of cleaning things up to
properly support replica sets. I have most of that working ina small
prototype locally, this is the process of shimming that into the
existing codebase.

This will take a few days, but seems to work well in my tests thus far.

Unit tests do not all pass, that will be fixed.