Fix reconnecting for _Connection class in txmongo.connection
Fix reconnecting. There is a bug when connection is lost, old (disconnected) proto is used, all deferreds that use this connection don't work properly and requests don't work after reconnnection at all. This is because getprotocol method of ConnectionPool return disconnected proto instance, which should be set to None on disconnection.