Code review and cleanup
Updates LICENSE, AUTHORS and CONTRIBUTORS files. 10gen is in the AUTHORS
list because we use ideas and some of their code from the original
pymongo driver. Christian have re/written large parts of the code.

The old README.rst is gone and we'll stick to README.md for now, which
is a mix of the two files.

Version has been updated to 0.5 so as debian package and rpm spec.