no more uncaught exception, add initial_time to connections in connection pool, use containers for travis-ci
drop py26 and pypy26 support, add pypy3 support.

try adding pypy3 explicitely

cryptography is ok with pypy 2.6, so adding that back... but travis only likes 2.4 so we are pushing this to allowed failures for now.

Exception.message was deprecated in 2.6, and removed in 3.0. Use str(Exception) instead.

add by pypy, but make sure it is ok to fail

see if we can get the latest pypy involved

pyflakes

allow failures

catch them all